KVM: nVMX: Use host_phys_bits for EPT tests that require 40+ PA bits
The EPT tests require at least 40 bits of physical address space so that
they can use a separate EPT L4 entry to hold guest data.  Qemu defaults
to 40 bits for the size of the guest PA, which causes the tests' check
against CPUID to always pass regardless of the system's actual PA space.
This results in unexpected reserved-bit page faults due to attempting to
access a GPA beyond the physical address width of the system, i.e. bit
39 is reserved.

Pass host_phys_bits for the tests to instruction Qemu to use the host's
number of physical address bits for the guest.  Qemu has supported this
option since v2.7, which was released in 2016, before the introduction
of the EPT tests in 2017.

GPA bits mapped to EPT levels:

    Offset: 11:0
    EPT L1: 20:12
    EPT L2: 29:21
    EPT L3: 38:30
    EPT L4: 47:39
    EPT L5: 56:48

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>