KVM: Add a "-t" option to run a specific test
x86's nested-VMX test space is massive, to the point where the group
option, "-g", does not provide sufficient granularity, e.g. a user may
want to run the core VMX tests but not the VMCS shadow tests, which take
an absurdly long time to run due to exhuastively testing all VMCS field
values.

Note that x86's config is currently a bit of a train wreck as far as VMX
is concerned; the "-t" option is of limited value until the config is
revamped.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>