x86: nVMX: Add tests of VMPTRLD and VMPTRST to MMIO addresses
KVM uses master abort semantics to handle MMIO accesses when emulating
VMX instructions.  Ensure this is the case, i.e. that KVM doesn't inject
a #PF, by emitting VMPTRLD and VMPTRST to the TPM address space.

Use the TXT private space as the bogus MMIO address so that the tests
can also pass on bare metal.  If my recollection of TXT is correct,
accesses to private space when it is locked trigger master aborts.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>