x86: access: Add test for illegal toggling of CR4.LA57 in 64-bit mode
Add a test to verify that KVM correctly injects a #GP if the guest
attempts to toggle CR4.LA57 while 64-bit mode is active.  Use two
versions of the toggling, one to toggle only LA57 and a second to toggle
PSE in additional to LA57.  KVM doesn't intercept LA57, i.e. toggling
only LA57 effectively tests the CPU, not KVM.  Use PSE as the whipping
boy as it will not trigger a #GP on its own, is universally available,
is ignored in 64-bit mode, and most importantly is trapped by KVM.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>