x86: apic: Play nice with x2APIC being enabled when getting "pre-boot" ID
Retrieve the "pre-boot" APIC ID via the x2APIC interface if x2APIC is
enabled instead of assuming that the APIC is always in xAPIC mode.   EFI
has a catch-22 where it needs to the APID ID to initialize the per-vCPU
GS.base, but calling reset_apic() = >disable_apic() needs GS.base to be
correctly initialized in order to set the correct APIC ops.  Play nice
with either xAPIC or x2APIC so that EFI can be used for SMP tests, and
in particular the SVM INIT-SIPI test which sends APs back through the
boot sequence while x2APIC is enabled.

Alternatively, disabling x2APIC and updating the APIC ops could be split,
but there's no obvious advantage in doing so.  Retrieving the pre-boot
APIC ID isn't a hot path, i.e. the cost of the RDMSR is likely neglibible,
and requiring callers to ensure the APIC is in the correct state is
unnecessarily fragile.

Signed-off-by: Sean Christopherson <seanjc@google.com>