nSVM: Add helper to report fatal errors in guest
Add a helper macro to dedup nSVM test code that handles fatal errors
by reporting the failure, setting the test stage to a magic number, and
invoking VMMCALL to bail to the host and terminate.

Note, the V_TPR fails if report() is invoked.  Punt on the issue for
now as most users already report only failures, but leave a TODO for
future developers.

Signed-off-by: Sean Christopherson <seanjc@google.com>