nVMX: Shuffle test_host_addr_size() tests to "restore" CR4 and RIP
Re-order the testcases in test_host_addr_size() to guarantee that the host
CR4 and RIP values are either written or restored before each testcase.
If a test fails unexpectedly, running with a test value from a previous
testcase may cause all subsequent tests to also fail, e.g. if the CR4
PCIDE test fails, all of the RIP tests will fail because of the bad CR4.

This also "fixes" the noncanonical RIP testcase, as running with the bad
CR4 setup by the !PAE testcase would mask a missed noncanonical check.

[sean: Surprise! The bad CR4 is indeed masking a bug.  I'm leaving it for
 now and intentionally creating a failing testcase for a commit or two to
 highlight the importance of cleaning up after testcases, and isolating
 what is actually being tested.]

Signed-off-by: Sean Christopherson <seanjc@google.com>