nVMX: Assert that the test is configured for 64-bit mode
The nVMX tests are 64-bit only, assert as much to reduce the probability
of the tests not actually running due to some funky bug.  It's easy to
re-add an if-statement if the nVMX tests ever gain 32-bit support, whereas
most developers lack the psychic powers necessary to detect that a test
isn't doing anything useful.

Signed-off-by: Sean Christopherson <seanjc@google.com>