x86/debug: Add a split-lock #AC / bus-lock #DB testcase
Add a testcase to the debug test to verify that a split-lock in the guest
does NOT result in an #AC or #DB (and that the test isn't killed).  While
KVM may run the guest with split-lock #AC enabled, KVM should never inject
an #AC into the guest.  And Bus Lock Detect/Trap should flat out never be
enabled while KVM is running the guest.

Link: https://bugzilla.kernel.org/show_bug.cgi?id=219787
Link: https://lore.kernel.org/all/bug-219787-28872@https.bugzilla.kernel.org%2F
Signed-off-by: Sean Christopherson <seanjc@google.com>