x86/debug: Add a testcase for DEBUGCTL.BTF, a.k.a. Block Single-Step
Add a testcase to verify Block Single-Step #DBs, i.e. single-stepping with
DEBUGCTL.BTF=1.  While KVM doesn't currently support BTF, KVM should at
least present consistent behavior to the guest.  Verify that KVM clears
BTF in the MSR, and that the guest sees normal single-step behavior.

If BTF is actually supported, verify it works according the x86
architecture, i.e. that #DBs occur on branch targets, and that BTF is
cleared on delivery of the #DB.

Not-yet-signed-off-by: Ravi Bangoria <ravi.bangoria@amd.com
[sean: use streamlined RFLAGS.TF sequence, play nice with KVM]
Signed-off-by: Sean Christopherson <seanjc@google.com>