x86/vmx: Track VMCS "launched" state per-CPU
Track whether or not a VMCS has been "launched" per-CPU to play nice with
VMX tests that use multiple VMCSes.  Arguably, it would be better to add a
structure to track the current VMCS and its launched state, but there's no
immediate benefit to doing so, and practically speaking the only thing that
would be tracked (without support for e.g. eVMCS) would be the "launched"
state.

Signed-off-by: Sean Christopherson <seanjc@google.com>