x86/kvmclock: Skip kvmclock test when not running on KVM with CLOCKSOURCE2
Skip the kvmclock test if the (virtual) CPU isn't running on KVM and/or
doesn't have CLOCKSOURCE2.  Presumably non-KVM environments simply don't
run the test, but checking for kvmclock support is easy enough.  E.g. with

  -cpu host,-kvmclock,kvm-pv-enforce-cpuid

the test will die on WRMSR #GPs without the checks, but generate

  SKIP: CPU not running on KVM with CLOCKSOURCE2
  SUMMARY: 1 tests, 1 skipped

with the appropriate checks.

Signed-off-by: Sean Christopherson <seanjc@google.com>