x86/smp: Align the stack to a 16-byte boundary when invoking SMP function calls
Align RSP to a 16-byte boundary in the IPI handler for SMP function calls
before calling into C code, as required by the x86-64 ABI.  Failure to
ensure the stack is properly aligned leads to obscure failures if a struct
(or any other object) tagged with __attribute__((aligned(16))) (or any
alignment greater than 16) is place on the stack.  E.g. VM-Enter will fail
on VMX if a vmx_msr_entry structure is placed on the stack.

Signed-off-by: Sean Christopherson <seanjc@google.com>