x86/vmx: Write to KVM's WALL_CLOCK MSR via VM-Entry load list sync in SIPI test
In the VMX Wait-for-SIPI => SIPI VM-Exit test, signal that the AP has
entered the guest by writing to MSR_KVM_WALL_CLOCK_NEW (when supported)
via the VM-Entry MSR load list instead of writing to memory from the AP
_before_ actually doing VM-Enter.  Abusing the MSR load list ensures that
the AP's "ready" signal to the BSP happens atomically with respect to
VM-Enter, and thus fixes a race where the BSP can see "ready" and send the
SIPI before the AP has executed VM-Enter.  E.g. with a delay inserted on
the AP, and no delay on the BSP, the test will hang 100% of the time.

Use MSR_KVM_WALL_CLOCK_NEW as it is pretty much the only MSR that KVM
emulates as a per-VM MSR, and that has a high likelihood of being
available.

Keep the BSP's delay before send the SIPI so that the test continues to
work if MSR_KVM_WALL_CLOCK_NEW isn't available, e.g. in bare metal (and
most KVM) setups, hitting the race is practically impossible.

Signed-off-by: Sean Christopherson <seanjc@google.com>